Move TokenVerifier to receiver package and TokenProvider to dispatcher package
