+++ data-plane/receiver/src/main/java/dev/knative/eventing/kafka/broker/receiver/impl/auth/TokenVerifierImpl.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.knative.eventing.kafka.broker.core.oidc;
+package dev.knative.eventing.kafka.broker.receiver.impl.auth;
 
 import dev.knative.eventing.kafka.broker.core.features.FeaturesConfig;
 import io.vertx.core.Future;
